backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Contract Reader dependencies
redis>=4.5.0
openai>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
Pillow>=10.0.0
PyPDF2>=3.0.0
pdfplumber>=0.9.0
pytesseract>=0.3.10
weasyprint>=60.0
pdf2image>=1.16.0

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.0.3

# AI and LLM
langchain==0.0.335
openai==1.3.5
tiktoken==0.7.0

# HTTP client for Jira API
httpx==0.25.2
aiohttp==3.9.1

# Security and authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Monitoring and observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-exporter-jaeger==1.21.0
prometheus-client==0.19.0

# Rate limiting
slowapi==0.1.9

# File processing and security
python-magic==0.4.27
clamd==1.0.2

# Configuration and utilities
python-dotenv==1.0.0
pyyaml==6.0.1
jinja2==3.1.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Logging
structlog==23.2.0
